Crate chardetng

chardetng is a character encoding detector for legacy Web content.

It is optimized for binary size in applications that already depend on encoding_rs for other reasons.

Modules

data 🔒

tld 🔒

Structs

ArabicFrenchCandidate 🔒

Big5Candidate 🔒

Candidate 🔒

CaselessCandidate 🔒

EncodingDetector

A Web browser-oriented detector for guessing what character encoding a stream of bytes is encoded in.

EucJpCandidate 🔒

EucKrCandidate 🔒

GbkCandidate 🔒

Iso2022Candidate 🔒

LatinCandidate 🔒

LogicalCandidate 🔒

NonLatinCasedCandidate 🔒

ShiftJisCandidate 🔒

Utf8Candidate 🔒

VisualCandidate 🔒

Enums

BeforeNonAscii 🔒

HalfWidthKatakana 🔒

InnerCandidate 🔒

Iso2022JpDetection

Whether to allow ISO-2022-JP as a guess result.

LatinCaseState 🔒

LatinCj 🔒

LatinKorean 🔒

NonLatinCaseState 🔒

OrdinalState 🔒

Utf8Detection

Whether to allow UTF-8 as a guess result.

Constants

BIG5_PUA_PENALTY 🔒

BIG5_SCORE_PER_LEVEL_1_HANZI 🔒

BIG5_SCORE_PER_OTHER_HANZI 🔒

BIG5_SINGLE_BYTE_EXTENSION_PENALTY 🔒

CJK_BASE_SCORE 🔒

CJK_LATIN_ADJACENCY_PENALTY 🔒

CJK_OTHER 🔒

CJK_SECONDARY_BASE_SCORE 🔒

CJ_PUNCTUATION 🔒

COPYRIGHT_BONUS 🔒

Must match the ISO-8859-2 score for “ Š “. Note: There are four Slovenian Wikipedia list page titles where the list is split by letter so that Š stands alone for the list part for Š. Let’s assume that’s a special case not worth detecting even though the copyright sign detection makes Slovenian title detection round to one percentage point worse.

EUC_JP_EXTENSION_PENALTY 🔒

EUC_JP_INITIAL_KANA_PENALTY 🔒

EUC_JP_SCORE_PER_KANA 🔒

EUC_JP_SCORE_PER_LEVEL_1_KANJI 🔒

EUC_JP_SCORE_PER_LEVEL_2_KANJI 🔒

EUC_JP_SCORE_PER_NEAR_OBSOLETE_KANA 🔒

EUC_JP_SCORE_PER_OTHER_KANJI 🔒

EUC_KR_HANJA_AFTER_HANGUL_PENALTY 🔒

EUC_KR_LONG_WORD_PENALTY 🔒

EUC_KR_MAC_KOREAN_PENALTY 🔒

EUC_KR_PUA_PENALTY 🔒

EUC_KR_SCORE_PER_EUC_HANGUL 🔒

EUC_KR_SCORE_PER_HANJA 🔒

EUC_KR_SCORE_PER_NON_EUC_HANGUL 🔒

EUC_KR_SINGLE_BYTE_EXTENSION_PENALTY 🔒

GBK_PUA_PENALTY 🔒

GBK_SCORE_PER_LEVEL_1 🔒

GBK_SCORE_PER_LEVEL_2 🔒

GBK_SCORE_PER_NON_EUC 🔒

GBK_SINGLE_BYTE_EXTENSION_PENALTY 🔒

HALF_WIDTH_KATAKANA_SCORE 🔒

HALF_WIDTH_KATAKANA_VOICING_SCORE 🔒

IMPLAUSIBILITY_PENALTY 🔒

IMPLAUSIBLE_LATIN_CASE_TRANSITION_PENALTY 🔒

LATIN_ADJACENCY_PENALTY 🔒

LATIN_LETTER 🔒

Latin letter caseless class

NON_LATIN_ALL_CAPS_PENALTY 🔒

NON_LATIN_CAPITALIZATION_BONUS 🔒

NON_LATIN_MIXED_CASE_PENALTY 🔒

ORDINAL_BONUS 🔒

SHIFT_JIS_EXTENSION_PENALTY 🔒

SHIFT_JIS_INITIAL_HALF_WIDTH_KATAKANA_PENALTY 🔒

SHIFT_JIS_PUA_PENALTY 🔒

SHIFT_JIS_SCORE_PER_KANA 🔒

SHIFT_JIS_SCORE_PER_LEVEL_1_KANJI 🔒

SHIFT_JIS_SCORE_PER_LEVEL_2_KANJI 🔒

SHIFT_JIS_SINGLE_BYTE_EXTENSION_PENALTY 🔒

Functions

cjk_extra_score 🔒

contains_upper_case_period_or_non_ascii 🔒

count_non_ascii 🔒

encoding_for_tld 🔒

encoding_is_native_to_tld 🔒

is_ascii_punctuation 🔒

more_problematic_lead 🔒

problematic_lead 🔒

score_adjustment 🔒