Crate chardetng

chardetng is a character encoding detector for legacy Web content.

It is optimized for binary size in applications that already depend on encoding_rs for other reasons.

Structs

EncodingDetector

A Web browser-oriented detector for guessing what character encoding a stream of bytes is encoded in.

Enums

Iso2022JpDetection

Whether to allow ISO-2022-JP as a guess result.

Utf8Detection

Whether to allow UTF-8 as a guess result.